[PATCH 319/350] close gather currently

+++ source/backend/tensorrt/execution/plugin/GatherPlugin.cpp
@@ -25,7 +25,7 @@ GatherPlugin::~GatherPlugin() {
 
 int GatherPlugin::onEnqueue(int batchSize, const void *const *inputs, void **outputs, void *, nvinfer1::DataType dataType, cudaStream_t stream) {
     const float *bottom_data = reinterpret_cast<const float *>(inputs[0]);
-    const int *indices = reinterpret_cast<const int *>(inputs[1]);
+    const float *indices = reinterpret_cast<const float *>(inputs[1]);
     float *top_data          = reinterpret_cast<float *>(outputs[0]);
     return GatherExecute(dataType, mCount, bottom_data, indices, top_data, stream);
 }
